fix input change, add constants, fix indents
