+++ public/main.html
@@ -29,6 +29,6 @@
     </main>
     <footer> </footer>
     <script type="module" src="./../src/scripts/main.js" async="async"></script>
-    <script type="module" src="./../src/scripts/APIController.js" async="async"></script>
+    <script type="module" src="./../src/scripts/APIController/APIController.js" async="async"></script>
   </body>
 </html>
