+++ public/search.html
@@ -43,5 +43,5 @@
     </main>
   </body>
   <script type="module" src="./../src/scripts/search.js" async="async"></script>
-  <script type="module" src="./../src/scripts/APIController.js" async="async"></script>
+  <script type="module" src="./../src/scripts/APIController/APIController.js" async="async"></script>
 </html>
