+++ src/scripts/Artist.js
@@ -1,14 +1,14 @@
 import  BaseView  from './BaseView.js';
+import {defaultViewImg} from './Constans.js'
 
 /** Сlass Artist */
 export default class Artist extends BaseView {
     name;
    
-
-      constructor(itemJson, idParent) {
+    constructor(itemJson, idParent) {
         itemJson.images.length
         ? super(itemJson.images[1].url, itemJson.id)
-        : super('artist.svg', itemJson.id); 
+        : super(defaultViewImg, itemJson.id); 
 
         this.name = itemJson.name;
