+++ src/scripts/Constans.js
@@ -1,0 +1,2 @@
+export const debounceTimeout = 500;
+export const defaultViewImg = 'artist.svg';
