+++ src/scripts/search.js
@@ -1,29 +1,27 @@
-import APIController from './APIController.js';
+import APIController from './ApiController/APIController.js';
 import Artist from './Artist.js';
 import {debounce} from './Helpers.js'
+import {debounceTimeout} from './Constans.js'
+import {urlSearch, clientId, clientSecret} from './ApiController/constants.js'
 
 
 (async function() {
-
-    const clientId = 'd0b55c9378d0473eb89d96c3b2e1a01c';
-    const clientSecret = '00f9a24f25254ee4a3313f66fdbb0968';
-
 
     /** Get data */ 
     const apiController = new APIController(clientId, clientSecret);
     const token = await apiController.getToken();
 
-    let inputSearch = document.querySelector('#search__input');
+    const inputSearch = document.querySelector('#search__input');
    
-    const onChangeDebounce = debounce(onChange, 500); 
+    const onChangeDebounce = debounce(onChange, debounceTimeout); 
     inputSearch.addEventListener('change', onChangeDebounce);
 
     /** 
      *  Search by query and insert search result in html
      */
-    async function onChange() {
+    async function onChange(event) {
         const searchResults = await apiController
-            .getData(apiController.urlSearch + inputSearch.value + '&type=artist', 0, token);
+            .getData(urlSearch + event.target.value + '&type=artist', token);
         
         if (searchResults.artists.items.length) {
             document.getElementById('results').replaceChildren();
